Add a space at end of PS1. Comment out source .bashrc
Sourcing .bashrc seems to be needed on interactive login. On Ubuntu desktop, .bash_profile is not run on login. For now, this can be worked around by sourcing it in .bashrc, but you can't have them source each other. Thus, the line sourcing .bashrc is left commented out for now.